requirements.txt
absl-py==1.3.0
appdirs==1.4.4
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.1
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1666764652625/work
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1671179356964/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
chex==0.1.5
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.6
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography-split_1677821248474/work
cycler==0.11.0
Cython==0.29.32
D4RL @ git+https://github.com/rail-berkeley/d4rl@71a9549f2091accff93eeff68f1f3ab2c0e0a288
decorator==5.1.1
distrax==0.1.3
dm-control==1.0.8
dm-env==1.5
dm-tree==0.1.7
docker-pycreds==0.4.0
etils==0.9.0
executing==1.2.0
fasteners==0.18
filelock==3.8.0
flax==0.6.1
flit_core @ file:///home/conda/feedstock_root/build_artifacts/flit-core_1667734568827/work/source/flit_core
fonttools==4.38.0
gast==0.5.3
gdown==4.5.3
gitdb==4.0.9
GitPython==3.1.31
glfw==2.5.5
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
gym==0.23.1
gym-notices==0.0.8
gym-robotics==0.1.0
h5py==3.8.0
icecream==2.1.3
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipdb==0.13.13
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
labmaze==1.0.5
lockfile==0.12.2
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-collections==0.1.1
mpi4py @ file:///home/conda/feedstock_root/build_artifacts/mpi4py_1636126593882/work
msgpack==1.0.4
mujoco==2.2.2
mujoco-py==2.1.2.14
multiworld==0.0.0
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
pandas==1.3.2
parso==0.8.3
patchelf==0.17.2.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1675487152289/work
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments==2.13.0
PyOpenGL==3.1.6
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1672659226110/work
pyparsing==2.4.7
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1673863902341/work
requests-oauthlib==1.3.1
-e git+ssh://git@github.com/hari-sikchi/ReCOIL_dev.git@c2704bfa24782ebcef65d1a104f977d3efc1ee2b#egg=research&subdirectory=online
rich==11.2.0
roboverse==0.0.0
rsa==4.9
scipy==1.9.3
sentry-sdk==1.10.1
setproctitle==1.3.2
shortuuid==1.0.9
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow-probability==0.18.0
tensorstore==0.1.28
termcolor==2.1.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
traitlets==5.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1673452138552/work
wandb==0.15.1
wcwidth==0.2.6
Werkzeug==2.2.2
zipp==3.10.0
